+++ src/cluster.rs
@@ -8,7 +8,7 @@ use crate::{Database, Error, schema::cluster::dsl, types::Cluster};
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct UpdateClusterRequest {
-    pub cluster_id: String,
+    pub cluster_id: i64,
     pub detector_id: i32,
     pub signature: String,
     pub score: Option<f64>,
@@ -21,7 +21,7 @@ pub struct UpdateClusterRequest {
 #[derive(Queryable)]
 struct ClusterDbSchema {
     id: i32,
-    cluster_id: String,
+    cluster_id: i64,
     category_id: i32,
     detector_id: i32,
     event_ids: Vec<Option<i64>>,
@@ -266,7 +266,7 @@ impl Database {
         Ok(())
     }
 
-    /// Find the numerical ids according to string ids of clusters for a model
+    /// Find the numerical ids according to numerical `cluster_ids` for a model
     ///
     /// # Errors
     ///
@@ -274,12 +274,12 @@ impl Database {
     pub async fn cluster_name_to_ids(
         &self,
         model_id: i32,
-        names: &[&str],
-    ) -> Result<Vec<(i32, String)>, Error> {
+        cluster_ids: &[i64],
+    ) -> Result<Vec<(i32, i64)>, Error> {
         let query = dsl::cluster
             .select((dsl::id, dsl::cluster_id))
             .filter(dsl::model_id.eq(&model_id))
-            .filter(dsl::cluster_id.eq_any(names));
+            .filter(dsl::cluster_id.eq_any(cluster_ids));
         let mut conn = self.pool.get().await?;
         Ok(query.get_results(&mut conn).await?)
     }
@@ -293,7 +293,7 @@ async fn upsert(
     use diesel::sql_types::{Array, BigInt, Double, Integer, Nullable, Text};
 
     let query = "SELECT attempt_cluster_upsert(
-        $1::text, $2::int4, $3::int8[], $4::text[], $5::int4, $6::text, $7::int8, $8::int4, $9::text[], $10::float8)";
+        $1::int8, $2::int4, $3::int8[], $4::text[], $5::int4, $6::text, $7::int8, $8::int4, $9::text[], $10::float8)";
     let (timestamps, sensors) =
         cluster
             .event_ids
@@ -305,7 +305,7 @@ async fn upsert(
             });
 
     diesel::sql_query(query)
-        .bind::<Text, _>(&cluster.cluster_id)
+        .bind::<BigInt, _>(&cluster.cluster_id)
         .bind::<Integer, _>(&cluster.detector_id)
         .bind::<Array<BigInt>, _>(&timestamps)
         .bind::<Array<Text>, _>(&sensors)